+++ src/redeneuraismlp/entidades/LinhaCSV.java
@@ -26,14 +26,14 @@ public class LinhaCSV {
         return atributos;
     }
     
-    public void setAtributo(String nome, float valor){
+    public void setAtributo(String nome, double valor){
         
         this.atributos.add(new Atributo(nome, valor));
     }
 
-    public List<Float> retornaValores(){
+    public List<Double> retornaValores(){
         
-        List<Float> valores = new ArrayList();
+        List<Double> valores = new ArrayList();
         
         for (int i = 0; i < atributos.size(); i++) {
